amplify/agent/requirements.txt
botocore
bedrock-agentcore
strands-agents
strands-agents[otel]
httpx
